+++ src/shared/api/base-api.ts
@@ -1,6 +1,6 @@
 import { createApi } from '@reduxjs/toolkit/query/react'
 
-import { customFetchBase } from '@/base-query-with-reauth'
+import { customFetchBase } from '@/features/auth/base-query-with-reauth'
 
 export const baseApi = createApi({
   reducerPath: 'baseApi',
