+++ src/shared/styles/globals.scss
